November 13th update
Small number of minor bug fixes whilst I await user feedback before continuing with this project.

+++ CosmicWatch v2 2021/CosmicWatch/CosmicWatch_Library/ReadFromFile.cs
@@ -35,7 +35,14 @@ namespace CosmicWatch_Library
         public bool Open(String Filename)
         {
             OpenFile = Path.Combine(new string[] { InputDirectory, Filename });
-            ReadFile = new StreamReader(OpenFile);
+            try
+            {
+                ReadFile = new StreamReader(OpenFile);
+            }
+            catch (Exception e)
+            {
+                ReadFile = null;
+            }
             return ReadFile != null;
         }
         public string ReadLine()
@@ -58,6 +65,7 @@ namespace CosmicWatch_Library
         public void DeleteFile()
         {
             ReadFile?.Close();
+            if (!File.Exists(OpenFile)) return;
             File.Delete(OpenFile);
             OpenFile = "";
         }
